+++ node/formats/eszip.ts
@@ -1,11 +1,12 @@
-import { join, resolve } from 'path'
-import { fileURLToPath } from 'url'
+import { join } from 'path'
 
+import type { WriteStage2Options } from '../../shared/stage2.js'
 import { DenoBridge } from '../bridge.js'
 import type { Bundle } from '../bundle.js'
 import { wrapBundleError } from '../bundle_error.js'
 import { EdgeFunction } from '../edge_function.js'
 import { ImportMap } from '../import_map.js'
+import { getPackagePath } from '../package_json.js'
 import { getFileHash } from '../utils/sha256.js'
 
 interface BundleESZIPOptions {
@@ -30,11 +31,11 @@ const bundleESZIP = async ({
   const extension = '.eszip'
   const destPath = join(distDirectory, `${buildID}${extension}`)
   const bundler = getESZIPBundler()
-  const payload = {
+  const payload: WriteStage2Options = {
     basePath,
     destPath,
     functions,
-    imports: importMap.imports,
+    importMapURL: importMap.toDataURL(),
   }
   const flags = ['--allow-all']
 
@@ -54,9 +55,8 @@ const bundleESZIP = async ({
 }
 
 const getESZIPBundler = () => {
-  const url = new URL(import.meta.url)
-  const pathname = fileURLToPath(url)
-  const bundlerPath = resolve(pathname, '../../../deno/bundle.ts')
+  const packagePath = getPackagePath()
+  const bundlerPath = join(packagePath, 'deno', 'bundle.ts')
 
   return bundlerPath
 }